[M90-LTS] Handle an empty tabstrip in TabStrip::GetDropBounds.
(cherry picked from commit 8c956ce0d96372940ec1702bbacec4b37cfa5357)

Bug: 1209616
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3017667
Commit-Queue: Taylor Bergquist <tbergquist@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#900744}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3085325
Owners-Override: Artem Sumaneev <asumaneev@google.com>
Commit-Queue: Roger Felipe Zanoni da Silva <rzanoni@google.com>
Cr-Commit-Position: refs/branch-heads/4430@{#1573}
Cr-Branched-From: e5ce7dc4f7518237b3d9bb93cccca35d25216cbe-refs/heads/master@{#857950}

+++ chrome/browser/ui/views/tabs/tab_strip.cc
@@ -3410,6 +3410,12 @@ gfx::Rect TabStrip::GetDropBounds(int drop_index,
                                   bool* is_beneath) {
   DCHECK_NE(drop_index, -1);
 
+  if (GetTabCount() == 0) {
+    // If the tabstrip is empty, it doesn't matter where the drop arrow goes.
+    // The tabstrip can only be transiently empty, e.g. during shutdown.
+    return gfx::Rect();
+  }
+
   Tab* tab = tab_at(std::min(drop_index, GetTabCount() - 1));
   int center_x = tab->x();
   const int width = tab->width();
